refactor: 클래스 import 방식 통일
