+++ src/class/index.js
@@ -1,2 +1,5 @@
 export { default as RacingGame } from './RacingGame.js';
+export { default as Cars } from './Cars.js';
+export { default as Car } from './Car.js';
+
 export { default as Validator } from './Validator.js';
